Correct some uses of 'URI' to 'URI reference' in structuring
While dealing with an issue in a set of schemas:
https://pagure.io/fedora-ci/messages/issue/67
I got rather confused by these docs until I went and read the
official spec drafts. In several places while discussing `$id`
and `$ref`, these docs claim these must be "URIs", but this is
not what the official draft says. It says they must be "URI-
references". From section 8.2, referring to `$id`:

"If present, the value for this keyword MUST be a string, and
MUST represent a valid URI-reference [RFC3986]."

From section 8.3, referring to `$ref`:

"The value of the "$ref" property MUST be a URI Reference."

This is a crucial and significant difference, as we can see in
RFC3986:

"A URI-reference is either a URI or a relative reference.  If
the URI-reference's prefix does not match the syntax of a scheme
followed by its colon separator, then the URI-reference is a
relative reference."

That is, there is no such thing as a "relative URI" (since RFC
3986 came in, anyway, and clarified this as compared to earlier
RFCs). URIs are absolute by definition. If the doc says something
must be a 'URI' then the most obvious reading is that it must be
an absolute-URI as defined in RFC3986. If it can in fact be a
relative reference, we must make this clear.

Signed-off-by: Adam Williamson <awilliam@redhat.com>