Cleanup and fixes for structuring
I didn't feel like this page presents the concepts very well. I ended up
rewriting the entire thing with the exception of the introduction and the
"Recursion" section. I think this revision does a better job at presenting the
concepts and is definitely more complete.

I wanted to include an example of how to use `$ref` and `definitions` to make
schemas more readable, but I'm having trouble making progress on that (I'm
foundering trying to come up with a real-ish example), so I'm leaving that out
for now.

Co-authored-by: about-code <6525873+about-code@users.noreply.github.com>
Co-authored-by: Ben Hutton <relequestual@gmail.com>