clarify that `type` for `additionalProperties` refer to the values
When `type` is specified in `additionalProperties`, such as in
the example below,

```
{
  "type": "object",
  "properties": {
    "number": { "type": "number" },
    "street_name": { "type": "string" },
    "street_type": { "enum": ["Street", "Avenue", "Boulevard"] }
  },
  "additionalProperties": { "type": "string" }
}
```

the doc currently says "... one can allow additional properties, but only
if they are each a string".

This is a little confusing, because "they" refers to the properties, and
it is unclear whether "they" refers to property keys or property values.

By changing it to "... one can allow additional properties, but only if
their values are each a string", this statement becomes unambiguous.